Develop

+++ workflow-processor/src/main/java/br/com/chain/workflow_processor/model/Profile.java
@@ -1,10 +1,5 @@
 package br.com.chain.workflow_processor.model;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-public record Profile(Long id, String name, String email) {
+public record Profile(int id, String name, String email) {
 
 }
